consumer/requirements.txt
python-dotenv
google-cloud-pubsub
psycopg2-binary
